ecourts-scraper/requirements.txt
requests
beautifulsoup4
python-dateutil
tqdm
pymupdf
flask
selenium
webdriver-manager
rich
pytesseract
Pillow
